# openapi.yaml
openapi: 3.0.3
info:
  title: MindWeaver API
  description: |
    MindWeaver 个人知识/研究动态监控工具 API

    功能特性：
    - RSS/Atom 订阅源管理
    - 条目获取、去重、存储和检索
    - 分类管理
    - 过滤规则管理
    - 自动调度获取
    - AI 摘要生成
    - 邮件摘要推送

  version: 1.0.0
  contact:
    name: MindWeaver
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/entries:
    get:
      tags: [Entries]
      summary: 获取所有条目
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: feed_id
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
          description: 搜索关键词
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
...
  /api/entries/batch/delete:
    post:
      tags: [Entries]
      summary: 批量删除条目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entry_ids
              properties:
                entry_ids:
                  type: array
                  items:
                    type: integer
                  description: 要删除的条目 ID 列表
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted_count:
                        type: integer
                  message:
                    type: string
  /api/entries/batch/fetch-content:
    post:
      tags: [Entries]
      summary: 批量获取完整内容
      description: 为指定条目获取完整文章内容
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entry_ids
              properties:
                entry_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: 操作完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      success:
                        type: integer
                        description: 成功数量
                      failed:
                        type: integer
                        description: 失败数量
                  message:
                    type: string
  /api/entries/batch/extract-keywords:
    post:
      tags: [Entries]
      summary: 批量提取关键词
      description: 为指定条目自动提取关键词
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entry_ids
              properties:
                entry_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: 操作完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      success:
                        type: integer
                      failed:
                        type: integer
                  message:
                    type: string
  /api/entries/batch/summarize:
    post:
      tags: [Entries]
      summary: 批量生成摘要
      description: 使用 AI 为指定条目生成摘要
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entry_ids
              properties:
                entry_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: 操作完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      success:
                        type: integer
                      failed:
                        type: integer
                  message:
                    type: string
  /api/entries/by-category/{category_id}:
    get:
      tags: [Entries]
      summary: 按分类获取条目
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      entries:
                        type: array
                        items:
                          $ref: '#/components/schemas/Entry'
                      total:
                        type: integer
                      page:
                        type: integer
                      page_size:
                        type: integer
  /api/entries/by-category-name/{category_name}:
    get:
      tags: [Entries]
      summary: 按分类名称获取条目
      parameters:
        - name: category_name
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      entries:
                        type: array
                        items:
                          $ref: '#/components/schemas/Entry'
                      page:
                        type: integer
                      page_size:
                        type: integer
  /api/entries/search-by-category/{category_id}:
    get:
      tags: [Entries]
      summary: 在分类中搜索条目
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      entries:
                        type: array
                        items:
                          $ref: '#/components/schemas/Entry'
                      query:
                        type: string
                      page:
                        type: integer
                      page_size:
                        type: integer
  /api/entries/by-category/{category_id}/stats:
    get:
      tags: [Entries]
      summary: 获取分类条目统计
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      category_id:
                        type: integer
                      category_name:
                        type: string
                      total_entries:
                        type: integer
                      recent_entries:
                        type: integer
                      top_sources:
                        type: array
                        items:
                          type: object

  # ==================== Scheduler API ====================
...
components:
  schemas:
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        feed_id:
          type: integer
          description: 所属订阅源 ID
        title:
          type: string
          description: 条目标题
        link:
          type: string
          format: uri
          description: 原始链接
        author:
          type: string
          description: 作者
        summary:
          type: string
          description: AI 生成的摘要
        content:
          type: string
          description: 完整内容
        published_at:
          type: string
          format: date-time
          description: 发布时间
        fetched_at:
          type: string
          format: date-time
          description: 获取时间
        tags:
          type: array
          items:
            type: string
          description: 标签列表
        language:
          type: string
          description: 语言代码
        reading_time_seconds:
          type: integer
          description: 预计阅读时间（秒）
        title_hash:
          type: string
          description: 标题哈希（用于去重）
        link_hash:
          type: string
          description: 链接哈希（用于去重）
        content_hash:
          type: string
          description: 内容哈希（用于去重）
        enabled:
          type: boolean
          description: 是否启用（未读=true，已读=false）
          default: true
...